Add helper method for creating packing plan
* Use helper method to dedup packing related code
* Move LauncherUtils to existing package
