+++ heron/scheduler-core/src/java/com/twitter/heron/scheduler/LaunchRunner.java
@@ -23,7 +23,6 @@ import com.twitter.heron.proto.system.ExecutionEnvironment;
 import com.twitter.heron.spi.common.Config;
 import com.twitter.heron.spi.common.Context;
 import com.twitter.heron.spi.common.Keys;
-import com.twitter.heron.spi.packing.IPacking;
 import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.scheduler.ILauncher;
 import com.twitter.heron.spi.statemgr.SchedulerStateManagerAdaptor;
@@ -39,14 +38,12 @@ public class LaunchRunner implements Callable<Boolean> {
   private Config runtime;
 
   private ILauncher launcher;
-  private IPacking packing;
 
   public LaunchRunner(Config config, Config runtime) {
 
     this.config = config;
     this.runtime = runtime;
     this.launcher = Runtime.launcherClassInstance(runtime);
-    this.packing = Runtime.packingClassInstance(runtime);
   }
 
   public ExecutionEnvironment.ExecutionState createExecutionState() {
@@ -116,9 +113,7 @@ public class LaunchRunner implements Callable<Boolean> {
     TopologyAPI.Topology topology = Runtime.topology(runtime);
     String topologyName = Context.topologyName(config);
 
-    // get the packed plan
-    packing.initialize(config, runtime);
-    PackingPlan packedPlan = packing.pack();
+    PackingPlan packedPlan = LauncherUtils.createPackingPlan(config, runtime);
     if (packedPlan == null) {
       LOG.severe("Failed to pack a valid PackingPlan. Check the config.");
       return false;
