+++ heron/schedulers/tests/java/com/twitter/heron/scheduler/aurora/AuroraLauncherTest.java
@@ -22,7 +22,7 @@ import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import com.twitter.heron.scheduler.common.LauncherUtils;
+import com.twitter.heron.scheduler.LauncherUtils;
 import com.twitter.heron.spi.common.Config;
 import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.scheduler.IScheduler;
